Code refactoring. The class and application code are in different files
